backend/requirements.txt
fastapi
fastapi[standard]
sqlalchemy
asyncpg
pydantic
pydantic-settings
supabase
alembic
python-multipart
email-validator
python-jose[crypt# filepath: c:\Code\System\Medora\backend\requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic
pydantic-settings
supabase
alembic
python-multipart
email-validator
python-jose[cryptography]
passlib[bcrypt]